+++ src/main/kotlin/com/dotori/v2/domain/member/presentation/AuthController.kt
@@ -3,11 +3,9 @@ package com.dotori.v2.domain.member.presentation
 import com.dotori.v2.domain.member.presentation.dto.req.SignInReqDto
 import com.dotori.v2.domain.member.presentation.dto.req.SignupReqDto
 import com.dotori.v2.domain.member.presentation.dto.res.SignInResDto
-import com.dotori.v2.domain.member.service.LogoutService
 import com.dotori.v2.domain.member.service.SignInService
 import com.dotori.v2.domain.member.service.SignupService
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -19,7 +17,6 @@ import javax.validation.Valid
 class AuthController(
     private val signupService: SignupService,
     private val signInService: SignInService,
-    private val logoutService: LogoutService,
 ) {
     @PostMapping("/signup")
     fun signup(@Valid @RequestBody signupReqDto: SignupReqDto): ResponseEntity<Void> =
@@ -29,9 +26,4 @@ class AuthController(
     @PostMapping
     fun signIn(@Valid @RequestBody signInReqDto: SignInReqDto): ResponseEntity<SignInResDto> =
         ResponseEntity.ok(signInService.execute(signInReqDto))
-
-    @DeleteMapping("/logout")
-    fun logout(): ResponseEntity<Void> =
-        logoutService.execute()
-            .run { ResponseEntity.ok().build() }
 }
